Add files via upload
Added user prompt function to BoxMaker (made in a function so it can be tested -- also can be passed a boolean that when false will not print out the user prompts themselves, this is useful so testing does not become crowded). The function simply returns true or false based on if Y or N is entered, and does not end until Y or N is entered. The goal of this function is to get used to user input so that prompt for dimensions later in the process is easier. The next commit will be the updated test file.