Add files via upload
Added another function on BoxMaker to get the SVG with specified length width and height values. Still in initial phase, the exact inches system hasn't been completely figured out yet and the thinkness of the wood is still not accounted for, so the pieces still won't fit together properly as of yet. Also, to keep existing tests running until future notice and changes, the original getSVG and static final string remain unchanged.