Add files via upload
Still without Christa's inserts SVG, this version added an additional function "getRelativeSVG" that will be called when the user specifies a dimension greater than 5in (new max of 25in in each direction). This creates a 55in x 85in (original dimensions times 5) that puts 0.5in in between each piece in each direction. If the normal getSVG is called with dimensions greater than 5, this one will automatically be called instead, making a large SVG with relative placement of the 6 sides of the box. With these additions all tests still pass.